requirements.txt
markdown2
watchdog
jinja2
python-frontmatter